yelp_data_analysis_demo/requirements/dev.txt
pytest==8.1.1
pytest-cov==5.0.0
isort==5.13.2
black==24.3.0
mypy==1.9.0
pylint==3.1.0
flake8==7.0.0
python-dotenv==1.0.1
pyspark==3.5.1
delta-spark==3.2.0
ipykernel